requirements.txt
pycryptodome==3.9.9
cryptography==3.4.7